[cluster-test] Move building methods from ClusterSwarm into ClusterBuilder
The goal of this change is actually to change Context::cluster_swarm from `ClusterSwarmKube` to `dyn ClusterSwarm` to make sure experiments only use API from ClusterSwarm interface.

But doing so requires removing implementation methods from ClusterSwarm trait, as otherwise it can not be made into object.
